Fix overflow when using //gm gil command
This wasn't the fault of the command or anything, but it turns out our
assumption of how quantities are stored client-side are wrong I guess.
Gil and other item quantities are actually 32-bit signed integers, so
they can only realistically hold 2147483647 in a stack.

I fixed the reported bug but there could be other places where I forgot
to change the quantity, I mainly focused on currencies.

Fixes #464

+++ servers/world/src/inventory/buyback.rs
@@ -55,7 +55,7 @@ impl BuyBackList {
         if !shop_buyback_items.is_empty() {
             for item in shop_buyback_items {
                 params[offset] = item.item_id;
-                params[offset + 1] = item.quantity;
+                params[offset + 1] = item.quantity as u32;
                 params[offset + 2] = item.price_low;
                 params[offset + 5] = shop_id;
                 params[offset + 8] = item.condition as u32;
